Fix : 모임 가입시 가입하지 않은 모임도 가입되었다고 뜨는 오류 수정

+++ src/routes/community.routes.js
@@ -77,9 +77,9 @@ router.delete(
 					.json({ message: '삭제하려는 모임이 존재하지 않습니다.' });
 			}
 
-			// if (loginId != existingCommunity.managerId) {
-			//   return res.status(403).json({ message: "모임장만 삭제 가능합니다." });
-			// }
+			if (loginId != existingCommunity.managerId) {
+			  return res.status(403).json({ message: "모임장만 삭제 가능합니다." });
+			}
 
 			await prisma.community.delete({
 				where: { id: +communityId },
@@ -116,6 +116,7 @@ router.post(
 		const User = await prisma.communityUsers.findFirst({
 			where: {
 				userId: +id,
+				communityId : +communityId,
 			},
 		});
 		if (User) {
